refactor: fix style
