+++ my-app/src/index.css
@@ -10,8 +10,10 @@ a {
   color: inherit;
 }
 
-body {
+body,
+html {
   background-color: #cfe8fc;
+  height: 100%;
 }
 
 .appBar.fixed {
@@ -20,15 +22,14 @@ body {
 }
 .appBar {
   transition: 0.8s ease-in;
-
 }
 
-/* html, 
+/* html,
 body {
   overflow-x: hidden;
 } */
 
- .box::-webkit-scrollbar-track {
+/* .box::-webkit-scrollbar-track {
   box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
   background-color: #f5f5f5;
 }
@@ -41,4 +42,4 @@ body {
 .box::-webkit-scrollbar-thumb {
   box-shadow: inset 0 0 6px hsla(0, 0%, 0%, 0.3);
   background-color: #555;
-} 
+} */
